Solution with useContext hook
By the use of useContext in the parent component (app.js), the whole page is not required to reload in order to pass the newly added name to the child component. (users.js)

Rather, here when a new name is added in the 'names' variable of the useContext the child component auto re-renders since it uses the variable in it's codeblock.